Update SQL schema and load scripts; enhance data types and add average rating query

+++ sql/sql_create/07_staging_schema.sql
@@ -12,7 +12,7 @@ CREATE TABLE staging_books (
 CREATE TABLE staging_users (
     user_id INT,
     location VARCHAR,
-    age FLOAT
+    age INT
 );
 
 CREATE TABLE staging_ratings (
