+++ sql/sql_load/06_load_users.sql
@@ -2,7 +2,7 @@ INSERT INTO users (id, location, age)
     SELECT
         user_id,
         location,
-        NULLIF(CAST(age AS FLOAT), 0)
+        NULLIF(CAST(age AS INT), 0)
     FROM
         staging_users
     WHERE
